app/requirements.txt
quart
quart-cors
telethon
python-dotenv
pyppeteer
hypercorn
httpx